requirements.txt
alembic==1.16.1
appdirs==1.4.4
APScheduler==3.11.0
babel==2.17.0
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.38.27
botocore==1.38.27
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
courlan==1.3.2
cssselect==1.3.0
dateparser==1.2.1
Events==0.5
faiss-cpu==1.11.0
fake-useragent==2.2.0
Flask==3.1.1
Flask-APScheduler==1.13.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
htmldate==1.9.3
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jusText==3.0.2
lxml==5.4.0
lxml_html_clean==0.4.2
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.2.6
opensearch-py==2.8.0
packaging==25.0
parse==1.20.2
pyee==11.1.1
PyMySQL==1.1.1
PyPDF2==3.0.1
pyppeteer==2.0.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
requests-aws4auth==1.3.1
requests-html==0.10.0
s3transfer==0.13.0
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.41
tiktoken==0.9.0
tld==0.13.1
tqdm==4.67.1
trafilatura==2.0.0
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==1.26.20
w3lib==2.3.1
websockets==10.4
Werkzeug==3.1.3
zipp==3.23.0
